requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
langchain==1.2.0
langchain-huggingface==1.2.0
langchain-groq==1.0.1
langchain-community==0.4.1
groq>=0.36.0
langchain-text-splitters==1.0.0
pypdf==6.4.1
